move source code into src/
