+++ src/rocrate_inveniordm/mapping/condition_functions.py
