+++ test/unit/test_converter.py
@@ -2,7 +2,7 @@ import re
 
 import pytest
 
-import mapping.converter as converter
+import rocrate_inveniordm.mapping.converter as converter
 
 
 def test_check_condition__true():
