+++ test/unit/test_processing.py
@@ -1,4 +1,4 @@
-import mapping.processing_functions as pf
+import rocrate_inveniordm.mapping.processing_functions as pf
 
 
 def test_dateProcessing__date_only():
